Update to portuguese artpieces
The edit was approved on discord after clarification of what happened with the scandal inside r/portugals discord group for r/place. The artists claim ownership for themselves and the new discord that took place, the description of this and more pieces will be edited to further show to the world what happened behind the scenes and the problems that streamers keep bringing to their country groups in r/place.